remove protos and openai key
