+++ example_config.toml
@@ -27,9 +27,5 @@ default_output_format = "text"
 # region = "eu1"
 # label = "production"
 #
-# Required for `cx search-fields`:
+# Required for `cx search-fields` and `cx metrics search --description`:
 # team_id = "123456"
-#
-# OpenAI API key used by `cx search-fields` for embedding generation.
-# The OPENAI_API_KEY environment variable takes precedence when set.
-# openai_api_key = "sk-..."
