+++ src/api/mod.rs
@@ -3,5 +3,5 @@ pub mod client;
 pub mod dashboards;
 pub mod dataprime;
 pub mod metrics;
-pub mod openai;
 pub mod schema_store;
+pub mod semantic_search;
