+++ src/commands/profiles.rs
@@ -289,12 +289,6 @@ async fn configure_oauth(name: &str) -> Result<(Profile, &'static str)> {
         Text::new("Coralogix team ID (required for search-fields):").prompt_skippable()?;
     let team_id = team_id.filter(|s| !s.is_empty());
 
-    let openai_api_key = Password::new("OpenAI API key (optional):")
-        .with_display_mode(PasswordDisplayMode::Masked)
-        .without_confirmation()
-        .prompt_skippable()?;
-    let openai_api_key = openai_api_key.filter(|s| !s.is_empty());
-
     // Clean up any existing secrets before writing new ones.
     keyring_store::delete_profile(name);
 
@@ -303,11 +297,8 @@ async fn configure_oauth(name: &str) -> Result<(Profile, &'static str)> {
     let tokens = oauth::browser_login(&base_url, &client_id).await?;
     println!("Login successful!");
 
-    // Store tokens and optional OpenAI key in the OS keyring.
+    // Store tokens in the OS keyring.
     oauth::store_tokens(name, &tokens)?;
-    if let Some(ref oai_key) = openai_api_key {
-        keyring_store::store_secret(name, "openai_api_key", oai_key)?;
-    }
 
     // For custom environments, explicitly store the base URL so that token
     // refresh can reach the correct OIDC discovery endpoint even if the
@@ -323,7 +314,6 @@ async fn configure_oauth(name: &str) -> Result<(Profile, &'static str)> {
         region,
         label,
         team_id,
-        openai_api_key: None,
         oauth_client_id: oauth_client_id_for_profile,
         oauth_base_url: oauth_base_url_for_profile,
     };
@@ -351,12 +341,6 @@ fn configure_api_key(name: &str) -> Result<(Profile, &'static str)> {
         Text::new("Coralogix team ID (required for search-fields):").prompt_skippable()?;
     let team_id = team_id.filter(|s| !s.is_empty());
 
-    let openai_api_key = Password::new("OpenAI API key (optional):")
-        .with_display_mode(PasswordDisplayMode::Masked)
-        .without_confirmation()
-        .prompt_skippable()?;
-    let openai_api_key = openai_api_key.filter(|s| !s.is_empty());
-
     let storage_choice = Select::new(
         "Where should API keys be stored?",
         CREDENTIAL_STORAGE_OPTIONS.to_vec(),
@@ -377,9 +361,6 @@ fn configure_api_key(name: &str) -> Result<(Profile, &'static str)> {
     let (profile, storage_desc) = match credential_storage {
         CredentialStorage::OsStore => {
             keyring_store::store_secret(name, "api_key", &api_key)?;
-            if let Some(ref oai_key) = openai_api_key {
-                keyring_store::store_secret(name, "openai_api_key", oai_key)?;
-            }
             let profile = Profile {
                 auth: AuthKind::ApiKey,
                 credential_storage,
@@ -387,7 +368,6 @@ fn configure_api_key(name: &str) -> Result<(Profile, &'static str)> {
                 region,
                 label,
                 team_id,
-                openai_api_key: None,
                 oauth_client_id: None,
                 oauth_base_url: None,
             };
@@ -403,7 +383,6 @@ fn configure_api_key(name: &str) -> Result<(Profile, &'static str)> {
                 region,
                 label,
                 team_id,
-                openai_api_key,
                 oauth_client_id: None,
                 oauth_base_url: None,
             };
