+++ src/commands/search_fields.rs
@@ -56,22 +56,7 @@ pub async fn run(
                      Run `cx profiles add` and enter your Coralogix team ID."
                 )
             })?;
-            let openai_key = t.cfg.openai_api_key.as_deref().ok_or_else(|| {
-                anyhow::anyhow!(
-                    "An OpenAI API key is required for search-fields.\n\
-                     Set OPENAI_API_KEY or run `cx profiles add` to store it in the profile."
-                )
-            })?;
-            semantic_field_lookup(
-                &t.cfg.endpoint,
-                &t.cfg.api_key,
-                team_id,
-                openai_key,
-                &tx,
-                &ds,
-                limit,
-            )
-            .await
+            semantic_field_lookup(&t.cfg.endpoint, &t.cfg.api_key, team_id, &tx, &ds, limit).await
         }
     })
     .await;
