+++ src/config.rs
@@ -209,7 +209,7 @@ mod max_size_serde {
 ///
 /// API keys are stored inline in the TOML by default (`credential_storage = "file"`).
 /// When `credential_storage = "os_store"`, keys are stored in the OS credential
-/// store and `api_key` / `openai_api_key` are `None` in the TOML.
+/// store and `api_key` is `None` in the TOML.
 ///
 /// OAuth profiles always use `credential_storage = "os_store"`; tokens are stored
 /// in the OS keyring and never written to the profile TOML.
@@ -231,10 +231,6 @@ pub struct Profile {
     /// Coralogix team ID, sent as `cgx-team-id` in gRPC metadata.
     #[serde(default)]
     pub team_id: Option<String>,
-    /// OpenAI API key for embedding-based features. Present when
-    /// `credential_storage = "file"`. Falls back to `OPENAI_API_KEY` env var.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub openai_api_key: Option<String>,
     /// OAuth: override the OAuth client ID.
     /// Required for custom regions; hard-coded for known Coralogix regions.
     /// Not written to the TOML for known regions (looked up from `KNOWN_ENVIRONMENTS`).
@@ -255,8 +251,6 @@ pub struct ResolvedConfig {
     pub endpoint: String,
     /// Coralogix team ID (`cgx-team-id` gRPC metadata).
     pub team_id: Option<String>,
-    /// OpenAI API key resolved from profile or `OPENAI_API_KEY` env var.
-    pub openai_api_key: Option<String>,
 }
 
 /// Returns the cx config directory: `~/.cx/`
@@ -315,8 +309,6 @@ pub fn load_profile(name: &str) -> Result<Profile> {
 ///   2. `AuthKind::ApiKey` — reads the key from OS keyring or profile file
 ///   3. `AuthKind::OAuth`  — loads the cached access token; refreshes via the
 ///      refresh token if the access token has expired (or is missing)
-///
-/// OpenAI key resolution: `OPENAI_API_KEY` env var > configured storage backend.
 async fn resolve_single(
     profile_name: &str,
     api_key_override: Option<&str>,
@@ -365,23 +357,11 @@ async fn resolve_single(
         }
     };
 
-    // OpenAI key: env var > storage backend
-    let openai_api_key = std::env::var("OPENAI_API_KEY")
-        .ok()
-        .filter(|s| !s.is_empty())
-        .or_else(|| match profile.credential_storage {
-            CredentialStorage::OsStore => keyring_store::get_secret(profile_name, "openai_api_key")
-                .ok()
-                .flatten(),
-            CredentialStorage::File => profile.openai_api_key.clone(),
-        });
-
     Ok(ResolvedConfig {
         profile_name: profile_name.to_string(),
         endpoint: profile.region.api_endpoint().to_string(),
         api_key: bearer,
         team_id: profile.team_id,
-        openai_api_key,
     })
 }
 
@@ -453,7 +433,6 @@ mod tests {
             api_key: "k".to_string(),
             endpoint: "https://api.eu2.coralogix.com".to_string(),
             team_id: None,
-            openai_api_key: None,
         };
         assert_eq!(cfg.profile_name, "prod");
     }
@@ -505,7 +484,6 @@ api_key = "mykey"
             region: Region::Eu2,
             label: Some("prod".to_string()),
             team_id: Some("12345".to_string()),
-            openai_api_key: None,
             oauth_client_id: None,
             oauth_base_url: None,
         };
@@ -528,7 +506,6 @@ api_key = "mykey"
             region: Region::Custom("https://api.myenv.coralogix.com".to_string()),
             label: None,
             team_id: None,
-            openai_api_key: None,
             oauth_client_id: Some("abc-123".to_string()),
             oauth_base_url: None,
         };
@@ -564,7 +541,6 @@ api_key = "mykey"
                 region: Region::Eu1,
                 label: None,
                 team_id: None,
-                openai_api_key: None,
                 oauth_client_id: None,
                 oauth_base_url: None,
             };
@@ -600,7 +576,6 @@ api_key = "mykey"
             region: Region::Eu1,
             label: None,
             team_id: None,
-            openai_api_key: None,
             oauth_client_id: None,
             oauth_base_url: None,
         };
@@ -622,7 +597,6 @@ api_key = "mykey"
             region: Region::Eu1,
             label: None,
             team_id: None,
-            openai_api_key: None,
             oauth_client_id: None,
             oauth_base_url: None,
         };
@@ -645,7 +619,6 @@ api_key = "mykey"
             region: Region::Eu1,
             label: None,
             team_id: None,
-            openai_api_key: None,
             oauth_client_id: None,
             oauth_base_url: None,
         };
@@ -668,7 +641,6 @@ api_key = "mykey"
             region: Region::Eu1,
             label: None,
             team_id: None,
-            openai_api_key: None,
             oauth_client_id: None,
             oauth_base_url: None,
         };
